feat: add reusable debounce hook for search
