+++ src/hooks/useDebounce.js
@@ -1,0 +1,12 @@
+import { useState, useEffect } from 'react';
+
+export default function useDebounce(value, delay = 300) {
+  const [debounced, setDebounced] = useState(value);
+
+  useEffect(() => {
+    const t = setTimeout(() => setDebounced(value), delay);
+    return () => clearTimeout(t);
+  }, [value, delay]);
+
+  return debounced;
+}
